requirements.txt
# Core deep learning
torch>=2.2.0
torchvision>=0.17.0

# Histology / WSI
openslide-python>=1.3.0
Pillow>=10.0.0

# Numerical / data
numpy>=1.26.0
scipy>=1.12.0
scikit-learn>=1.4.0
pandas>=2.0.0

# Visualisation
matplotlib>=3.8.0
opencv-python>=4.9.0

# Experiment tracking (optional but recommended)
wandb>=0.16.0

# Demo (optional)
gradio>=4.0.0
